@@ -3,27 +3,18 @@ description: A simple command-line application.
 # version: 1.0.0
 # homepage: https://www.example.com
 environment:
-  sdk: ">=2.12.0-259.12.beta <3.0.0"
+  sdk: ">=2.13.4 <3.0.0"
 
 dev_dependencies:
-  build_runner: null
-  pedantic: ^1.9.0
+  build_runner: ^2.1.0
+  pedantic: ^1.11.1
 dependencies:
-  args: ^1.6.0
-  dotenv: ^2.0.0
-  get_it: ^5.0.6
-  html: ^0.15.0
-  json_serializable: ^4.0.0
-  logging: ^1.0.0-nullsafety.0
-  mongo_dart: ^0.4.4
-  nyxx: 1.1.0-dev.3
-  nyxx_commander: 1.1.0-dev.3
-  nyxx_interactions:
-    git:
-      url: https://github.com/MattheousDT/nyxx.git
-      ref: bugfix/commandarg-json-encode-error
-      path: nyxx.interactions
-dependency_overrides:
-  logging: ^1.0.0-nullsafety.0
-  http: ^0.12.0
-  json_annotation: ^4.0.0
+  args: ^2.2.0
+  dotenv: ^3.0.0
+  get_it: ^7.2.0
+  json_serializable: ^5.0.0
+  logging: ^1.0.1
+  mongo_dart: ^0.7.1
+  nyxx: ^2.0.0-rc.12
+  nyxx_interactions: ^2.0.0-rc.17
+  path: ^1.8.0
