@@ -1,18 +1,16 @@
 import 'package:logging/logging.dart';
 import 'package:mongo_dart/mongo_dart.dart';
 
-import '../models/user.dart';
+import '../models/user/user.dart';
 
 class UserService {
-  final Logger _logger = Logger('NOMAC | UserService');
+  UserService(Db db) : _coll = db.collection('users');
 
-  final Db db;
-  final String collectionName = 'users';
-
-  UserService(this.db);
+  final _logger = Logger('NOMAC | UserService');
+  final DbCollection _coll;
 
   Future<NomacUser?> getUserByDiscordId(String discordId) async {
-    Map<String, dynamic>? result = await db.collection(collectionName).findOne(where.eq('discord_id', discordId));
+    final result = await _coll.findOne(where.eq('discord_id', discordId));
 
     if (result == null) {
       return null;
@@ -23,7 +21,7 @@ class UserService {
 
   Future<bool> createUser(NomacUser user) async {
     try {
-      await db.collection(collectionName).insert(user.toJson());
+      await _coll.insert(user.toJson());
       return true;
     } catch (err) {
       _logger.severe(err);
@@ -32,26 +30,24 @@ class UserService {
   }
 
   Future<NomacUser> updateUser(NomacUser user) async {
-    var coll = db.collection(collectionName);
-
-    Map<String, dynamic> result = await coll.findOne(where.eq('discord_id', user.discordId));
+    var result = await _coll.findOne(where.eq('discord_id', user.discordId));
 
     if (result == null) {
       await createUser(user);
-      result = await coll.findOne(where.eq('discord_id', user.discordId));
+      result = await _coll.findOne(where.eq('discord_id', user.discordId));
     }
 
     user.toJson().entries.forEach((element) {
-      result[element.key] = element.value;
+      result![element.key] = element.value;
     });
 
-    await coll.save(result);
+    await _coll.replaceOne(where.eq('_id', result!['_id']), result);
     return user;
   }
 
   Future<bool> deleteUser(NomacUser user) async {
     try {
-      await db.collection(collectionName).remove(where.eq('discord_id', user.discordId));
+      await _coll.remove(where.eq('discord_id', user.discordId));
       return true;
     } catch (err) {
       _logger.severe(err);
